Extract data for `embedding-snippets` spec into a separate file
