+++ frontend/test/metabase/scenarios/embedding/embedding-snippets.cy.spec.js
@@ -1,5 +1,7 @@
 import { restore, popover, visitDashboard } from "__support__/e2e/cypress";
 
+import { JS_CODE, IFRAME_CODE } from "./embedding-snippets";
+
 describe("scenarios > embedding > code snippets", () => {
   beforeEach(() => {
     restore();
@@ -16,38 +18,6 @@ describe("scenarios > embedding > code snippets", () => {
     cy.findByText(
       "Insert this code snippet in your server code to generate the signed embedding URL",
     );
-
-    const JS_CODE = new RegExp(
-      `// you will need to install via 'npm install jsonwebtoken' or in your package.json
-
-var jwt = require("jsonwebtoken");
-
-var METABASE_SITE_URL = "http://localhost:PORTPORTPORT";
-var METABASE_SECRET_KEY = "KEYKEYKEY";
-var payload = {
-  resource: { dashboard: 1 },
-  params: {},
-  exp: Math.round(Date.now() / 1000) + (10 * 60) // 10 minute expiration
-};
-var token = jwt.sign(payload, METABASE_SECRET_KEY);
-
-var iframeUrl = METABASE_SITE_URL + "/embed/dashboard/" + token + "#bordered=true&titled=true";`
-        .split("\n")
-        .join("")
-        .replace(/[-\/\\^$*+?.()|[\]{}]/g, "\\$&")
-        .replace("KEYKEYKEY", ".*")
-        .replace("PORTPORTPORT", ".*"),
-    );
-
-    const IFRAME_CODE = `<iframe
-    src="{{iframeUrl}}"
-    frameborder="0"
-    width="800"
-    height="600"
-    allowtransparency
-></iframe>`
-      .split("\n")
-      .join("");
 
     cy.get(".ace_content")
       .first()
